cell_manufacturing_project/requirements.txt
blinker==1.9.0
click==8.1.7
clickhouse-driver==0.2.9
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kafka-python==2.0.2
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.2
tzlocal==5.2
Werkzeug==3.1.3
